Calculate single quantile for numerical values

Slightly faster than quantile with na.rm=TRUE. The internal implementation uses the 'C++' function std::nth_element, which is significantly faster than base R implementation when the length of input x is less than 1e7.

Usage

fastquantile(x, q)

Arguments

x

numerical vector (integers or double)

q

number from 0 to 1

Value

Identical to quantile(x, q, na.rm=TRUE)

Examples

# create input x with NAs
x <- rnorm(10000)
x[sample(10000, 10)] <- NA

# compute median
res <- fastquantile(x, 0.5)
res
#> [1] 0.01298609

# base method
res == quantile(x, 0.5, na.rm = TRUE)
#>  50% 
#> TRUE 
res == median(x, na.rm = TRUE)
#> [1] TRUE

# Comparison
microbenchmark::microbenchmark(
  {
    fastquantile(x, 0.5)
  },{
    quantile(x, 0.5, na.rm = TRUE)
  },{
    median(x, na.rm = TRUE)
  }
)
#> Unit: microseconds
#>                                    expr     min       lq     mean  median
#>            {     fastquantile(x, 0.5) }  83.917 116.9225 131.5374 125.489
#>  {     quantile(x, 0.5, na.rm = TRUE) } 287.917 299.4530 356.6948 312.277
#>         {     median(x, na.rm = TRUE) } 216.554 229.1125 280.9078 237.513
#>        uq     max neval
#>  148.2965 195.665   100
#>  436.1530 475.887   100
#>  361.0680 407.690   100