Create R object map.

Provides five types of map that fit in different use cases.

session_map(map = fastmap::fastmap())

rds_map(path = tempfile())

text_map(path = tempfile())

qs_map(path = tempfile())

redis_map(name = rand_string())

Arguments

map
a fastmap::fastmap() list

path
directory path where map data should be stored

name
character, map name. If map names are the same, the data will be shared.

Value

An R6 instance that inherits AbstractMap

Details

There are five types of map implemented. They all inherit class AbstractMap. There are several differences in use case scenarios and they backend implementations.

session_map

A session map takes a fastmap object. All objects are stored in current R session. This means you cannot access the map from other process nor parent process. The goal of this map is to share the data across different environments and to store global variables, as long as they share the same map object. If you are looking for maps that can be shared by different processes, check the rest map types. The closest map type is redis_map, which is also memory-based.

rds_map

An 'RDS' map uses file system to store values. The values are stored separately in '.rds' files. Compared to session maps, 'RDS' map can be shared across different R process. It's recommended to store large files in rds_map. If the value is not large in RAM, text_map and redis_map are recommended.

qs_map

A 'qs' map uses package 'qs' as backend. This map is very similar to rds_map, but is especially designed for large values. For example, pushing 1GB data to qs_map will be 100 times faster than using rds_map, and text_map will almost fail. However, compared to rds_map the stored data cannot be normally read by R as they are compressed binary files. And qs_map is heavier than text_map.

text_map

A 'text' map uses file system to store values. Similar to rds_map, it can be stored across multiple processes as long as the maps share the same file directory. However, unlike rds_map, text_map the text_map can only store basic data values, namely atom data types. The supported types are: numeric, character, vector, list, matrix It's highly recommended to convert factors to characters. Do NOT use if the values are functions or environments. Please check write_yaml for details. The recommended use case scenario is when the speed is not the major concern, and you want to preserve data with backward compatibility. Otherwise it's highly recommended to use redis_map, qs_map, and rds_map.

redis_map

A 'Redis' map uses free open source software `Redis` and R package 'RcppRedis' as backend. Compared to session map, 'Redis' map can be shared across sessions. Compared to 'text' and 'rds' maps, 'Redis' map stores data in memory, meaning a potential of significant speed ups. To use redis_map, you need to install `Redis` on your computer.

On Mac: use `brew install redis` to install and `brew services start redis` to start the service

On Linux: use `sudo apt-get install redis-server` to install and `sudo systemctl enable redis-server.service` to start the service

On Windows: Download from https://github.com/dmajkic/redis/downloads and double click 'redis-server.exe'

Examples

# ----------------------Basic Usage ----------------------

# Define a path to your map.
path = tempfile()
map <- qs_map(path)

# Reset
map$reset()

# Check if the map is corrupted.
map$validate()

# You have not set any key-value pairs yet.
# Let's say two parallel processes (A and B) are sharing this map.
# Process A set values
map$keys()

#> NULL

# Start push
# set a normal message
map$set(key = 'a', value = 1)

# set a large object
map$set(key = 'b', value = rnorm(100000))

# set an object with hash of another object
map$set(key = 'c', value = 2, signature = list(
  parameter1 = 123,
  parameter2 = 124
))

# Check what's in the map from process B
mapB <- qs_map(path)
mapB$keys()

#> ==YQ ==Yg ==Yw 
#>  "a"  "b"  "c" 

mapB$keys(include_signatures = TRUE)

#>      keys                                   
#> ==YQ "a"  "6717f2823d3202449301145073ab8719"
#> ==Yg "b"  "ccbaa6726fd79ecdc62206d33eed8109"
#> ==Yw "c"  "f3b07dfebb13e24c3b1a5e11ec21a15d"

# Number of key-values pairs in the map.
mapB$size()

#> [1] 3

# Check if key exists
mapB$has(c('1','a', 'c'))

#>     1     a     c 
#> FALSE  TRUE  TRUE 

# Check if key exists and signature also matches
mapB$has('c', signature = list(
  parameter1 = 123,
  parameter2 = 124
))

#>    c 
#> TRUE 

# Signature changed, then return FALSE. This is especially useful when
# value is really large and reading the value takes tons of time
mapB$has('c', signature = list(
  parameter1 = 1244444,
  parameter2 = 124
))

#>     c 
#> FALSE 

# Destroy the map's files altogether.
mapB$destroy()

if (FALSE) {
  # Once destroyed, validate will raise error
  mapB$validate()
}